Test JsonDocument instead of JsonSerializer
